arm-trusted-firmware: add patches from solidrun repo
These patches will work on the official cn9130-cf-base, but not our
prototype unit. In order to enable compilation for our prototype unit a
hacky append has been added just to keep things clean and not add a new
target just for a single unit.

This append can be activated by placing a symlink to it.